requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymongo==4.6.2
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.33.0
pydantic==2.6.4